equip gun ui
added ui element to show when and which gun can be equipped. gun equipping is now done with additional overlapping collider on player instead of line trace. equippable guns is now an array to deal with multiple guns in range. picked up one is currently the one at 0 index (i.e. the first encountered)