Equipping animation
gun assets have no equipping animations or anything similar, and i'm not an animator myself, so i've "animated" then just with translations and rotations.
when a gun is equipped (picked up or unholstered), a timeline is used to change it's position over time as a sort of equipping animation. this animation's time can be changed, and will differ per gun.